+++ src/bandcamp/modules/html.js
@@ -46,6 +46,6 @@ export function getBandPhotoSrc() {
 }
 
 export function isValidBandcampURL(url) {
-  const pattern = /^(https?:\/\/)?([a-z0-9-]+\.)*bandcamp\.com(\/[a-z0-9-]+)*(\/[a-z0-9-]+\/[a-z0-9-]+)?$/;
-  return pattern.test(url);
+  const bandcampUrl = 'bandcamp.com';
+  return url.includes(bandcampUrl);
 }