+++ src/discogs/app/utils.js
@@ -6,3 +6,8 @@
 export const getDiscogsDateValue = date => {
   return date.toISOString().split('T')[0];
 };
+
+export const isValidDiscogsReleaseEditUrl = (url) => {
+  const discogsEditUrl = 'www.discogs.com/release/edit/';
+  return url.includes(discogsEditUrl);
+};