+++ src/discogs/content.js
@@ -1,22 +1,11 @@
 'use strict';
 
-injectCSSFile(chrome.runtime.getURL('src/discogs/notification.css'));
-injectJSFile(chrome.runtime.getURL('src/discogs/script.js'), () => { console.log('B2D: Discogs script loaded'); });
+importModule('src/discogs/modules/content-main.js');
 
-/**
- * @param {String} cssUrl
- */
-function injectCSSFile(cssUrl) {
-  const linkElement = document.createElement('link');
-  linkElement.rel = 'stylesheet';
-  linkElement.href = cssUrl;
-
-  document.head.appendChild(linkElement);
-}
-
-function injectJSFile(url, callback) {
-  const s = document.createElement('script');
-  s.src = url;
-  s.onload = callback;
-  (document.head||document.documentElement).appendChild(s);
+function importModule(url) {
+  (async () => {
+    const src = chrome.runtime.getURL(url);
+    const contentMain = await import(src);
+    contentMain.main();
+  })();
 }