+++ src/popup/tabs/history_tab.js
@@ -1,21 +1,41 @@
 import { DiscogsCsv } from "../../discogs/app/discogs-csv.js";
+import { ReleasesList } from '../components/releases-list.js';
 import { downloadCsv, objectsToCsv } from "../../modules/csv.js";
 import { createElementFromHTML, hasDataAttribute, setDataAttribute } from "../../modules/html.js";
-import { clearStorage, clearStorageByKey, findAllReleases, findReleasesByUrls, logStorage } from "../../modules/storage.js";
+import { clearStorage, clearStorageByKey, findAllReleases, findReleasesByUrls } from "../../modules/storage.js";
 import { populateReleasesList, removeButtonLoadingState, setButtonInLoadingState } from "../helpers.js";
+import config from "../../config.js";
+import { loadDiscogsGenres } from "../../discogs/modules/genres.js";
+import { loadKeywordMapping } from "../../bandcamp/modules/mapping.js";
 
 /**
  * @param {Element} btnDownloadCsv
  */
 export function setupHistoryTab(tab, btnDownloadCsv) {
-  const releasesList = document.querySelector('#historyReleasesList');
+  const releasesList = getReleasesList();
 
   if (!hasDataAttribute(tab, 'buttons-initialized')) {
     setupReleasesList(tab, releasesList, btnDownloadCsv);
     setDataAttribute(tab, 'buttons-initialized');
   }
 
-  updateReleasesListData(releasesList);
+  loadDiscogsGenres(config.genres_url).then(genres => {
+    loadKeywordMapping(config.keyword_mapping_url).then(keywordsMapping => {
+      updateReleasesListData(releasesList);
+    })
+  });
+}
+
+export function setHistoryTabSearchValue(searchValue) {
+  const releasesList = getReleasesList();
+  releasesList.setSearchValue(searchValue);
+}
+
+/**
+ * @returns {ReleasesList}
+ */
+function getReleasesList() {
+  return document.getElementById('historyReleasesList');
 }
 
 function updateReleasesListData(releasesList) {